cifar-10/requirements.txt
future
numpy
opencv
scikit-image
scikit-learn
Pillow
EasyDict
azure-storage
